+++ src/pages/Impressum/Impressum.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Box } from '../../components/Base/Base';
-import TextAdress from './TextAdress';
-import PrivacyPolicy from './PrivacyPolicy';
+import TextAdress from './components/TextAdress';
+import TextPrivacyPolicy from './components/TextPrivacyPolicy';
 
 export default function Impressum() {
     return (
         <Box lineHeight="1.3rem">
             <TextAdress />
-            <PrivacyPolicy />
+            <TextPrivacyPolicy />
         </Box>
     );
 }
