+++ src/pages/Landingpage/Landingpage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Flex } from '../../components/Base/Base';
-import ImageMaximilian from './ImageMaximilian';
-import TextWelcome from './TextWelcome';
+import ImageMaximilian from './components/ImageMaximilian';
+import TextWelcome from './components/TextWelcome';
 
 export default function Landingpage() {
     return (
